requirements.txt
beautifulsoup4==4.9.3
lxml==4.6.3
python-dotenv==0.16.0
requests==2.25.1
terminaltables==3.1.0
